Working multiplayer test server
Server with implemented gameloop that supports multiple connected players

Listens after incoming packets, when a packet is received it is added to a packet queue. A packetparsing thread then picks the packet from 
the queue and parses data contained in the packet. parsing is done by making a new player object with a x and y coordinate and adding 
this object to the playerlist. 

Everytime the gameloop runs (every 16 ms) a packet is sendt to registered players that contains name, x and y coordinates for every registered
players.

Also added comments to every class and methods.

Next step is to add the network implementation to the game and add functionality

git-svn-id: http://www.nlogn.org/svn/epsilon/trunk@110 189e0cc2-7543-4f94-9523-152a149c9a55