Working "thin" multiplayer test client
Listens after incoming packets, when a packet is received it is added to a packet queue. A packetparsing thread then picks the packet from 
the queue and parses data contained in the packet. parsing is done by making a new player object with a x and y coordinate and adding 
this object to the playerlist. when incoming gamestate is parsed a call is done to repaint() and all player entities are drawn on the testpanel. 

Everytime the mouse is moved over the testpanel a packet is sent to the server containing playername, x and y coordinates. 

Also added comments to every class and methods.

Next step is to add the network implementation to the game and add functionality

git-svn-id: http://www.nlogn.org/svn/epsilon/trunk@111 189e0cc2-7543-4f94-9523-152a149c9a55