Added network subsystem to the game
Network subsystem includes a thread for listening after incoming packets, 
a thread for parsing incoming packets and a thread for sending gamestate to the server.
It also includes a singleton that acts as a interface to the network subsystem.

TODO: the parsing thread is not yet complete, and we should add handshake functionality to establish a connection to the server.

git-svn-id: http://www.nlogn.org/svn/epsilon/trunk@136 189e0cc2-7543-4f94-9523-152a149c9a55